% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvforest.train.R
\name{lvforest.train}
\alias{lvforest.train}
\title{Estimating Valid Latent Variable Scores from Conditionally Causal Models}
\usage{
lvforest.train(
  data,
  model,
  input,
  ntrees = 100,
  split = 2,
  minsize = 300,
  cutoff_rmsea = 0.05,
  cutoff_loading = 0.2,
  dbsamp = T,
  bagging = NULL,
  ordered = NULL,
  std.lv = FALSE,
  ctree_control = partykit::ctree_control(minbucket = minsize, mtry = split)
)
}
\arguments{
\item{data}{Dataset to be analyzed. Note that the observed responses of your model cannot contain missing data.}

\item{model}{Model description in lavaan-terminology.}

\item{input}{Character vector of partitioning variables. Note that every partitioning variable must either be defined as factor (if it should be treated as categorical) or numeric.}

\item{ntrees}{Number of trees to be computed. Default = 100.}

\item{split}{Number of partitioning variables to be selected for random split selection at every split point within a tree ('mtry' argument in ctree function). Default = 2.}

\item{minsize}{Minimum size of terminal nodes of trees. Needs to be large enough to estimate the model parameters. Default = 300.}

\item{cutoff_rmsea}{Cutoff value for model fit evaluation. Default = .05.}

\item{cutoff_loading}{Cutoff value for multiplicative parameters in model. Note that models with parameter estimates not significantly different form 0 are excluded per default. Default = 0.}

\item{dbsamp}{Double sampling. Default = TRUE.}

\item{bagging}{Numeric vector defining the proportion of original data used for tree growing (first number in vector) and re-fitting terminal nodes (second number in vector). Default = NULL.}

\item{ordered}{Character vector of ordinal observed variables of the model. Default = NULL.}

\item{std.lv}{Standardization of latent variable variances in model. Default = FALSE.}

\item{ctree_control}{Control settings for ctree function. Default = ctree_control(minbucket = minsize, mtry = split).}
}
\description{
Estimating Valid Latent Variable Scores from Conditionally Causal Models
}
\examples{
\dontrun{
trained <- lvforest.train(
input = c("conf","dicho1","cat1","num1","cat2","rand1","rand2","rand3","rand4","rand5"),
model <- 'LatVar1 =~ simuvar1 + beta2*simuvar2 + beta3*simuvar3 
 LatVar2 =~ simuvar4 + beta5*simuvar5 + beta6*simuvar6
 LatVar3 =~ simuvar7 + beta8*simuvar8 + beta9*simuvar9
 LatVar4 =~ LatVar1 + lambda2*LatVar2 + lambda3*LatVar3 + delta*simuvar_effect',
ordered = c("simuvar1","simuvar2","simuvar3","simuvar4",
"simuvar5","simuvar6","simuvar7","simuvar8","simuvar9"),
data = simu,
cutoff_rmsea = .03,
ctree_control = ctree_control(minbucket = 200, mtry = 1, testtype = "Teststatistic") 
#Bonferroni correction off 
)
}
}
