% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scforest.predict.R
\name{scforest.predict}
\alias{scforest.predict}
\title{Estimating Valid Latent Variable Scores from Conditionally Causal Models}
\usage{
scforest.predict(
  trained,
  data,
  idvar,
  conf = trained$Info$input,
  latvar = latvars[1],
  indmethod = "gamma",
  exclude_unconf = TRUE,
  stdscores = FALSE
)
}
\arguments{
\item{trained}{Trained model created by scforest.train().}

\item{data}{Dataset to be used for latent variable score prediction. Note that the observed responses of your model cannot contain missing data and that it must contain all partitioning variables from training.}

\item{idvar}{ID-Variable to distinguish individual data points in the dataset.}

\item{conf}{Character vector of potential confounding variables. Default = 'Partitioning Variables from training'.}

\item{latvar}{Latent variable of interest to estimate latent variable scores.}

\item{indmethod}{Method for testing for significance of dHSIC. Default = gamma.}

\item{exclude_unconf}{Only use latent variable scores if there is unconfoundedness. Default = TRUE.}

\item{stdscores}{Latent variable scores as z-scores.}
}
\description{
Estimating Valid Latent Variable Scores from Conditionally Causal Models
}
\examples{
\dontrun{
predicted <- scforest.predict(trained = trained,
data = simu, idvar = "id", latvar = "LatVar4", exclude_unconf = T, stdscores = F)
}
}
